# backend/internal/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Lynx Server API
  version: 0.1.0
  description: Local Lynx Guardian backend API served by the Gin-based lynx-server binary.
servers:
  - url: http://127.0.0.1:31789
    description: Default local lynx-server
...
paths:
...
  /lynx/protected-resources:
    post:
      tags: [query]
      operationId: upsertProtectedResource
      summary: Create or update one protected resource policy.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonObject"
      responses:
        "200":
          $ref: "#/components/responses/ObjectResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /lynx/policy-rules:
    post:
      tags: [query]
      operationId: upsertPolicyRule
      summary: Create or update one user-managed policy rule.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonObject"
      responses:
        "200":
          $ref: "#/components/responses/ObjectResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
...
    ObjectResponse:
      description: Single JSON object result.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JsonObject"
    ErrorResponse:
      description: Error payload.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    JsonObject:
      type: object
      additionalProperties: true
...
    ErrorResponse:
      type: object
      required: [ok, message]
      properties:
        ok:
          type: boolean
          enum: [false]
        message:
          type: string
...